+++ R/corrp2.R
@@ -1,0 +1,9 @@
+#' @inherit corrp title params return seealso example
+#' @description same function as \code{corrp}.
+#' @export
+corrP <-
+function(df, ...) {
+
+  UseMethod("corrp",df)
+
+}