+++ CHANGELOG.md
@@ -4,4 +4,6 @@
 
 - Change package name `corrP` to `corrp` ;
 - Changelog file created ;
-- Fixed some bugs in function's documentations.
+- Fixed some bugs in function's documentations;
+- Add verbose param to corrp function ; 
+-